[#42] Entity 파일 분리
